+++ tests/testthat/test-utils.R
@@ -17,3 +17,20 @@ test_that("sort_files works", {
      tolower(files_sorted)
    )
 })
+
+test_that("print utilities work", {
+
+   expect_snapshot(pui_done("wooo!"))
+   expect_snapshot(pui_info("so, ..."))
+   expect_snapshot(pui_oops("well, ..."))
+   expect_snapshot(pui_todo("next, ..."))
+
+   # make sure this runs quiet
+   withr::local_options(projthis.quiet = TRUE)
+
+   expect_snapshot(pui_done("wooo!"))
+   expect_snapshot(pui_info("so, ..."))
+   expect_snapshot(pui_oops("well, ..."))
+   expect_snapshot(pui_todo("next, ..."))
+
+})